Avoid starting a drag event when the deselecting
When the user wants to (de)select a project/feature/metadata then
sometimes the press gets slightly off. This can happen when the force on
the mouse or touchpad is not optimal and can occur for any user.

Only consider a drag event when the distance traveled during the click
is more than 20% of the average width of the selected element.