Scatterplot: Consider scaling in tooltip distance
The calculation of the nearest point for a tooltip location used the
domain of the actual values, which causes strange behavior when one of
the axes has a very large scale (such as number of lines of code) while
the other does not. Use the conversion function to calculate the
distance but keep the output values as is for further processing by the
tooltip focus handler.