Chart: Try to keep tooltip within axis
This alters the tooltip positioning near the axis edges such that it
should prefer to remain within them. In some charts the overlay
width/height may extend over the chart axis (for easier brushing) but
this still enables tooltips to not overlap with them. In other charts
this ensures the tooltip items remain clickable in the overlay.