Expand sprint source URLs to use correct data
For charts, the source URL should not always link to the latest URL, and
we now support partially expanded source URL which we render during the
visualization. Also, we support team projects with multiple project keys
but a singular board, and merged sprints, to provide the URL with the
correct (number of) IDs.