Chart: Use short units for stacked assignments if available
If a feature that is used in an expression that is shown in a stacked
assignment (i.e. in the bar chart), then use its short unit if it has
one. Fallback to longer units if there are no short units. This shortens
the tooltip text since the assignment itself should already have some
descriptive text so the short unit usually suffices.