Filter features in configuration to relevant ones
This filters the features based on groups that the formatter considers
relevant. For now, this only has an effect for the Sankey chart format
which requires features that have details.

Selected features are not hidden. CSS is adjusted to keep the features
look well. The clear button no longer removes selected projects by
correcting the URL upon any updates.