Format: Link domain names to correct source type
If a source ID has a different source type than sonar, then use
a different link generator to link to the domain object at the source.

Currently only Git is supported as an alternative source type.

Find source description based on feature source type match: If there are
multiple sources for a domain object then use the one that is applicable
for the source that the feature has.