Adjust sprint selection to consider future sprints
Future sprints are now given negative ticks in the selection based on
their recentness. This is propagated to the data selection which adjusts
its filtering to remove future/old sprints based on the position from
the most recent active sprint.

The sprint selection hides the option to select future sprints if the
formatter does not use them or the features do not have predictions.

The ticks of the sprint selection may now contain another tick between
the minimum/maximum value (number of future/current sprints) and the
tick before it, to avoid large parts of emptyness in the selection axis
but still avoiding overlapping the ticks through digit counting.